Fix PHPStan failure: keep the GF Text test stub out of static analysis
The inline Text stand-in in DownloadUrlTest.php shadowed the real
Gravity_Forms\...\Settings\Fields\Text class for PHPStan (which scans
vendor/gravityforms), producing 7 errors across DownloadUrl.php and the
test. Move the stub to tests/stubs/ (excluded from analysis, loaded from
the PHPUnit bootstrap), store the $settings argument on it so the double
also supports DownloadUrl::save(), pass a Mockery Settings mock in the
tests to satisfy the real constructor's docblock type, and register the
already-installed phpstan-mockery extension for mock type inference.
